Split out tabs from modal.
