+++ packages/interface/src/style.scss
@@ -5,5 +5,6 @@
 @import "./components/more-menu-dropdown/style.scss";
 @import "./components/pinned-items/style.scss";
 @import "./components/preferences-modal/style.scss";
+@import "./components/preferences-modal-tabs/style.scss";
 @import "./components/preferences-modal-section/style.scss";
 @import "./components/preferences-modal-base-option/style.scss";
